pyproject.toml
[project]
name = "beermonopoly-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "blessed>=1.20.0",
    "cloudscraper>=1.2.71",
    "cryptography>=43.0.3",
    "django-admin-shell>=2.0.1",
    "django-allauth>=65.1.0",
    "django-anymail>=12.0",
    "django-cors-headers>=4.5.0",
    "django-dirtyfields>=1.9.3",
    "django-extensions>=3.2.3",
    "django-filter>=24.3",
    "django-hosts>=6.0",
    "django-q2>=1.7.3",
    "django>=4.2.16",
    "djangorestframework>=3.15.2",
    "drf-dynamic-fields>=0.4.0",
    "filters>=1.3.2",
    "freezegun>=1.5.1",
    "fuzzywuzzy>=0.18.0",
    "googlesearch-python>=1.2.5",
    "gunicorn>=23.0.0",
    "jupyter>=1.1.1",
    "lxml>=5.3.0",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.9.0",
    "pytest-django>=4.9.0",
    "pytest>=8.3.3",
    "responses>=0.25.3",
    "ruff>=0.12.0",
    "sentry-sdk>=2.17.0",
    "urllib3>=2.2.3",
    "xmltodict>=0.14.2",
]

[tool.uv]
dev-dependencies = [
    "codecov>=2.1.13",
    "coverage>=7.6.4",
]
